+++ company/conftest.py
@@ -7,39 +7,6 @@ import pytest
 
 
 @pytest.fixture
-def retrieve_profile_data():
-    return {
-        'website': 'http://example.com',
-        'description': 'Ecommerce website',
-        'summary': 'this is a short summary',
-        'number': '01234567',
-        'sectors': ['SECURITY'],
-        'logo': 'nice.jpg',
-        'name': 'Great company',
-        'slug': 'great-company',
-        'keywords': 'word1, word2',
-        'employees': '501-1000',
-        'date_of_creation': '2015-03-02',
-        'verified_with_code': True,
-        'twitter_url': 'http://www.twitter.com',
-        'facebook_url': 'http://www.facebook.com',
-        'linkedin_url': 'http://www.linkedin.com',
-        'supplier_case_studies': [],
-        'modified': '2016-11-23T11:21:10.977518Z',
-        'email_full_name': 'Jeremy',
-        'email_address': 'test@example.com',
-        'postal_full_name': 'Jeremy',
-        'address_line_1': '123 Fake Street',
-        'address_line_2': 'Fakeville',
-        'locality': 'London',
-        'postal_code': 'E14 6XK',
-        'po_box': 'abc',
-        'country': 'GB',
-        'mobile_number': '07506043448',
-    }
-
-
-@pytest.fixture
 def list_public_profiles_data(retrieve_profile_data):
     return {
         'results': [
