+++ conftest.py
@@ -6,6 +6,62 @@ import requests
 import pytest
 
 
+@pytest.fixture()
+def captcha_stub():
+    # https://github.com/praekelt/django-recaptcha#id5
+    os.environ['RECAPTCHA_TESTING'] = 'True'
+    yield 'PASSED'
+    os.environ['RECAPTCHA_TESTING'] = 'False'
+
+
+@pytest.fixture
+def retrieve_profile_data():
+    return {
+        'website': 'http://example.com',
+        'description': 'Ecommerce website',
+        'summary': 'this is a short summary',
+        'number': '01234567',
+        'sectors': ['SECURITY'],
+        'logo': 'nice.jpg',
+        'name': 'Great company',
+        'slug': 'great-company',
+        'keywords': 'word1, word2',
+        'employees': '501-1000',
+        'date_of_creation': '2015-03-02',
+        'verified_with_code': True,
+        'twitter_url': 'http://www.twitter.com',
+        'facebook_url': 'http://www.facebook.com',
+        'linkedin_url': 'http://www.linkedin.com',
+        'supplier_case_studies': [],
+        'modified': '2016-11-23T11:21:10.977518Z',
+        'email_full_name': 'Jeremy',
+        'email_address': 'test@example.com',
+        'postal_full_name': 'Jeremy',
+        'address_line_1': '123 Fake Street',
+        'address_line_2': 'Fakeville',
+        'locality': 'London',
+        'postal_code': 'E14 6XK',
+        'po_box': 'abc',
+        'country': 'GB',
+        'mobile_number': '07506043448',
+    }
+
+
+@pytest.fixture
+def search_results(retrieve_profile_data):
+    return {
+        'hits': {
+            'total': 1,
+            'hits': [
+                {
+                    '_source': retrieve_profile_data
+
+                }
+            ]
+        }
+    }
+
+
 @pytest.fixture
 def api_response_200():
     response = requests.Response()
@@ -14,9 +70,14 @@ def api_response_200():
     return response
 
 
-@pytest.fixture()
-def captcha_stub():
-    # https://github.com/praekelt/django-recaptcha#id5
-    os.environ['RECAPTCHA_TESTING'] = 'True'
-    yield 'PASSED'
-    os.environ['RECAPTCHA_TESTING'] = 'False'
+@pytest.fixture
+def api_response_search_200(api_response_200, search_results):
+    api_response_200.json = lambda: search_results
+    return api_response_200
+
+
+@pytest.fixture
+def api_response_400():
+    response = requests.Response()
+    response.status_code = 400
+    return response
