+++ exportopportunity/helpers.py
@@ -1,15 +1,30 @@
 from api_client import api_client
 
-from company.helpers import format_company_details
+from company.helpers import format_case_study, format_company_details
+
+
+def get_showcase_resource(sector, api_client_method, formatter):
+    response = api_client_method(
+        term="", page=1, sectors=sector, size=3
+    )
+    response.raise_for_status()
+    formatted = []
+    for result in response.json()['hits']['hits']:
+        formatted.append(formatter(result['_source']))
+    return formatted
 
 
 def get_showcase_companies(sector):
-    response = api_client.company.search(
-        term="", page=1, sectors=sector, size=3,
+    return get_showcase_resource(
+        sector=sector,
+        api_client_method=api_client.company.search_company,
+        formatter=format_company_details
     )
-    response.raise_for_status()
-    parsed = response.json()
-    formatted = []
-    for result in parsed['hits']['hits']:
-        formatted.append(format_company_details(result['_source']))
-    return formatted
+
+
+def get_showcase_case_studies(sector):
+    return get_showcase_resource(
+        sector=sector,
+        api_client_method=api_client.company.search_case_study,
+        formatter=format_case_study
+    )
