+++ exportopportunity/templates/exportopportunity/campaign-base.html
@@ -116,9 +116,9 @@
                         {% block carousel_slide_one_button_text %}{% endblock %}
                     </a>
                 </div>
-                <div class="ed-carousel__slide__image span6" style="background-image: url('{{ case_studies.0.logo }}');">
+                <div class="ed-carousel__slide__image span6" style="background-image: url('{{ case_studies.0.image }}');">
                     <div class="ed-carousel__slide__image_caption">
-                        {{ case_studies.0.name }} | <a href="{{ case_studies.0.case_study_url }}">view case study</a>
+                        {{ case_studies.0.title }} | <a href="{{ case_studies.0.case_study_url }}">view case study</a>
                     </div>
                 </div>
               </div>
@@ -133,9 +133,9 @@
                         {% block carousel_slide_two_button_text %}{% endblock %}
                     </a>
                 </div>
-                <div class="ed-carousel__slide__image span6" style="background-image: url('{{ case_studies.1.logo }}');">
+                <div class="ed-carousel__slide__image span6" style="background-image: url('{{ case_studies.1.image }}');">
                     <div class="ed-carousel__slide__image_caption">
-                        {{ case_studies.1.name }} | <a href="{{ case_studies.1.case_study_url }}">view case study</a>
+                        {{ case_studies.1.title }} | <a href="{{ case_studies.1.case_study_url }}">view case study</a>
                     </div>
                 </div>
               </div>
@@ -150,9 +150,9 @@
                         {% block carousel_slide_three_button_text %}{% endblock %}
                     </a>
                 </div>
-                <div class="ed-carousel__slide__image span6" style="background-image: url('{{ case_studies.2.logo }}');">
+                <div class="ed-carousel__slide__image span6" style="background-image: url('{{ case_studies.2.image }}');">
                     <div class="ed-carousel__slide__image_caption">
-                        {{ case_studies.2.name }} | <a href="{{ case_studies.2.case_study_url }}">view case study</a>
+                        {{ case_studies.2.title }} | <a href="{{ case_studies.2.case_study_url }}">view case study</a>
                     </div>
                 </div>
               </div>
